starting em-components.ts
